Improve docker error handling
Use case when docker service is not running won't be treated as error.
The package is refactored to simplify the implementation. The files renamed for consistency.
