+++ pkg/monitoring/docker/docker_windows.go
@@ -2,16 +2,10 @@
 
 package docker
 
-type Watcher struct{}
-
-func New() (*Watcher, error) {
+func ListContainers() (map[string]interface{}, error) {
 	return nil, ErrorNotImplementedForOS
 }
 
-func (*Watcher) ListContainers() (map[string]interface{}, error) {
-	return nil, ErrorNotImplementedForOS
-}
-
-func (*Watcher) ContainerNameByID(_ string) (string, error) {
+func ContainerNameByID(_ string) (string, error) {
 	return "", ErrorNotImplementedForOS
 }
